+++ src/net/fexcraft/app/fmt/ui/general/Button.java
@@ -97,7 +97,7 @@ public class Button extends Element {
 	}
 	
 	@Override
-	public void hovered(int mx, int my){
+	public void hovered(float mx, float my){
 		super.hovered(mx, my); if(this.hovered){ for(Element elm : elements) elm.setVisible(true); }
 	}
